+++ example/lib/main.dart
@@ -9,13 +9,8 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      // title: 'Flutter_Custom_Cards Demo',
-      title: 'Custom Card Example',
+      title: 'CustomCards',
       debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-      ),
       home: CustomCardsExample(),
     );
   }
@@ -27,93 +22,132 @@ class CustomCardsExample extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         centerTitle: true,
-        title: Text(
-          // 'Flutter_Custom_Cards Example',
-          'Custom Card Example',
-        ),
+        title: Text('Custom Cards Example'),
       ),
-      body: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Column(
-          children: [
-            Text('TextCards'),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                //TextCard Exmple
-                TextCard(
-                  text: 'Card1',
-                  childPadding: 12,
-                  fontSize: 28,
+      body: SingleChildScrollView(
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
+            children: [
+              Text('CustomCards'),
+              Center(
+                child: Wrap(
+                  //mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    CustomCard(
+                      height: 50,
+                      child: FlutterLogo(
+                        style: FlutterLogoStyle.horizontal,
+                        size: 90,
+                      ),
+                    ),
+                    CustomCard(
+                      height: 50,
+                      width: 100,
+                      elevation: 6,
+                      childPadding: 10,
+                      color: Colors.green,
+                      onTap: () {},
+                      child: Center(
+                        child: Text(
+                          'Flutter',
+                          style: TextStyle(
+                            fontSize: 21,
+                            color: Colors.white,
+                          ),
+                        ),
+                      ),
+                    ),
+                    CustomCard(
+                      height: 50,
+                      width: 100,
+                      borderRadius: 10,
+                      color: Colors.red,
+                      hoverColor: Colors.indigo,
+                      splashColor: Colors.white,
+                      onTap: () {},
+                      child: Center(
+                        child: Text(
+                          'Flutter',
+                          style: TextStyle(
+                            fontSize: 20,
+                            color: Colors.white,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
                 ),
-                TextCard(
-                  text: 'Card2',
-                  childPadding: 12,
-                  fontSize: 28,
-                  color: Colors.green,
-                  textColor: Colors.white,
-                ),
-                TextCard(
-                  text: 'Card3',
-                  childPadding: 12,
-                  fontSize: 28,
-                  color: Colors.red,
-                  textColor: Colors.white,
-                ),
-              ],
-            ),
-            SizedBox(
-              height: 20,
-            ),
-            Text('ImageCards'),
-            Row(
-              children: [
-                //ImageCard Example
-                ImageCard(
-                  elevation: 12,
-                  childPadding: 10,
-                  shadowColor: Colors.green,
-                  image: Image.network(
-                    'https://miro.medium.com/max/85/1*ilC2Aqp5sZd1wi0CopD1Hw.png',
-                  ),
-                ),
-                ImageCard(
-                  elevation: 8,
-                  shadowColor: Colors.red,
-                  childPadding: 10,
-                  color: Colors.yellow,
-                  image: Image.network(
-                    'https://miro.medium.com/max/85/1*ilC2Aqp5sZd1wi0CopD1Hw.png',
-                  ),
-                ),
-                ImageCard(
-                  elevation: 6,
-                  childPadding: 10,
-                  color: Colors.purple,
-                  image: Image.network(
-                    'https://miro.medium.com/max/85/1*ilC2Aqp5sZd1wi0CopD1Hw.png',
-                  ),
-                ),
-              ],
-            ),
-            SizedBox(
-              height: 20,
-            ),
-            Text('Widget Card'),
-            WidgetCard(
-              widget: Column(
+              ),
+              SizedBox(height: 8),
+              Wrap(
+                //mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  Image.network(
-                      'https://miro.medium.com/max/85/1*ilC2Aqp5sZd1wi0CopD1Hw.png'),
-                  Text('Widget Card Example'),
-                  ElevatedButton(
-                    onPressed: () {},
-                    child: Text('Button'),
+                  CustomCard(
+                    height: 100,
+                    width: 100,
+                    elevation: 5,
+                  ),
+                  CustomCard(
+                    height: 100,
+                    width: 100,
+                    elevation: 5,
+                    borderRadius: 50,
+                    color: Colors.green,
+                    hoverColor: Colors.yellow,
+                    onTap: () {},
+                  ),
+                  CustomCard(
+                    height: 100,
+                    width: 100,
+                    elevation: 5,
+                    color: Colors.blue,
+                    borderColor: Colors.white,
+                    borderWidth: 2,
                   ),
                 ],
               ),
-            ),
-          ],
+              SizedBox(height: 25),
+              Text('Custom3DCards'),
+              Wrap(
+                //mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.all(15.0),
+                    child: Custom3DCard(
+                      elevation: 4,
+                      child: SizedBox(
+                        height: 100,
+                        width: 100,
+                        child: Center(
+                          child: FlutterLogo(size: 65),
+                        ),
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.all(15.0),
+                    child: Custom3DCard(
+                      elevation: 10,
+                      shadowSpread: 5,
+                      shadowColor: Colors.brown.shade400,
+                      child: SizedBox(
+                        height: 100,
+                        width: 100,
+                        child: Center(
+                          child: FlutterLogo(
+                            size: 65,
+                            style: FlutterLogoStyle.stacked,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+              SizedBox(height: 30),
+            ],
+          ),
         ),
       ),
     );