+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.0] - 20/11/2021
+
+This is a big update (major changes)
+
+The following **major changes** happened
+- `TextCard`, `ImageCard` and `WidgetCard` have been merged into single card named `CustomCard`.
+- Now new CustomCards supports onTap, hoverColor, height, width and more.
+- `Custom3DCard` is introduced in this version that was not in the previous version.
+- Now there are only two cards `CustomCard` and `Custom3DCard`
+
 ## [0.1.0] - 25/03/2021
 
 - Migrated to `null-safety`